Issues with Certain Counties
Hello reader,

I was facing some issues with certain counties with geopy. So in order to try fix it I created a program to find certain line from my address row and if it matches I add the county into it. The only issue you might have is you have to change the ,, to , manually. You can do this by doin CTRL + F and use Replace All in Visual Studio. Hope this helps :D.